Tims changes but reverting back to PSimHitMap
